+++ site/scenes/index.js
@@ -6,10 +6,11 @@
 //
 // Adding an animation is two edits: drop a scene module in this folder, then add it here.
 
+import * as roseFunnel from './rose-funnel/index.js';
 import * as aboveTheFog from './above-the-fog/index.js';
 import * as grizzlyPeak from './grizzly-peak/index.js';
 import * as floatingBed from './floating-bed/index.js';
 
-export const scenes = [aboveTheFog, grizzlyPeak, floatingBed];
+export const scenes = [roseFunnel, aboveTheFog, grizzlyPeak, floatingBed];
 
 export const findScene = (id) => scenes.find((scene) => scene.meta.id === id);